+++ server/clients/stellar/account.go
@@ -3,7 +3,6 @@ package stellargoclient
 import (
 	"bytes"
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 
 	"github.com/pkg/errors"
@@ -12,69 +11,6 @@ import (
 	"github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/txnbuild"
 )
-
-func (c *Client) getTFTTransactionFundingCondition() (feeWallet, fee string, err error) {
-	baseUrl := c.GetTransactionFundingUrlFromNetwork()
-	resp, err := http.Get(baseUrl + "/conditions")
-	if err != nil {
-		return
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	type Condition struct {
-		Asset          string
-		Fee_account_id string
-		Fee_fixed      string
-	}
-	conditions := make([]Condition, 0)
-	json.Unmarshal(body, &conditions)
-	tftAsset := c.GetTftAsset()
-	tftAssetString := tftAsset.GetCode() + ":" + tftAsset.GetIssuer()
-	for _, c := range conditions {
-		if c.Asset == tftAssetString {
-			feeWallet = c.Fee_account_id
-			fee = c.Fee_fixed
-			return
-		}
-	}
-	err = errors.New("Transaction funding condition for TFT not found")
-	return
-}
-
-// Allows you to fund a transaction with tft instead of lumen by using the transaction funding service: https://github.com/threefoldfoundation/tft-stellar/tree/master/ThreeBotPackages/transactionfunding_service
-func (c *Client) FundTransactionUsingTft(destination string, amount string) error {
-	feeAccount, fee, err := c.getTFTTransactionFundingCondition()
-	if err != nil {
-		return errors.Wrap(err, "failed to get tft transaction funding condition")
-	}
-	payment := txnbuild.Payment{
-		Destination: destination,
-		Amount:      amount,
-		Asset:       c.GetTftAsset(),
-	}
-	feePayment := txnbuild.Payment{
-		Destination: feeAccount,
-		Amount:      fee,
-		Asset:       c.GetTftAsset(),
-	}
-
-	sourceAccount, err := c.AccountData(c.kp.Address())
-	if err != nil {
-		return errors.Wrap(err, "failed to get accountdetail")
-	}
-	tx, err := txnbuild.NewTransaction(txnbuild.TransactionParams{
-		SourceAccount:        &sourceAccount,
-		IncrementSequenceNum: true,
-		BaseFee:              0,
-		Operations:           []txnbuild.Operation{&payment, &feePayment},
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewInfiniteTimeout(),
-		},
-	})
-	if err != nil {
-		return errors.Wrap(err, "failed to create transaction")
-	}
-	return c.SignFundAndSubmitTransaction(tx)
-}
 
 // Activates the account using the activation service https://github.com/threefoldfoundation/tft-stellar/tree/master/ThreeBotPackages/activation_service
 func (c *Client) activateAccount() error {
